Make AddAsync and AddRangeAsync include related entities (1.1.1 fix)
Issue #7298

This is a lower-risk fix replacing #7430. The fix is to call the non-async methods. There is an extremely low probability that any real-world code is making database calls in the async methods combined with an even lower probability that making those calls synchronously would cause issues, so this fix is low-risk and does not have binary-compat issues.